+++ components/Catalog/CatalogBoxMod.js
@@ -3,6 +3,7 @@ import { StyleSheet, View, Image, Text, TouchableOpacity } from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
 
 const ITEM_SIZE = 160;
+// const test = 3
 
 class CatalogBoxMod extends Component {
   render() {
@@ -10,8 +11,16 @@ class CatalogBoxMod extends Component {
       <View style={styles.container}>
         <Image source={{ uri: this.props.product_image }} style={styles.image} />
         <LinearGradient colors={['rgba(0, 0, 0, 0.8)', 'transparent']} style={styles.gradient}>
-          <Text style={styles.prodName}>{this.props.product_name}</Text>
-          <View style={styles.whitePill}><Text style={styles.whitePillText}>{'RM '+(this.props.product_price).toFixed(2)}</Text></View>
+        {this.props.product_name!=undefined?
+        <Text style={styles.prodName}>{this.props.product_name}</Text>:
+        <Text style={styles.prodName}>loading..</Text>
+        }
+        {this.props.product_price!=null?
+        <View style={styles.whitePill}><Text style={styles.whitePillText}>{'RM '+ Number(this.props.product_price).toFixed(2)}</Text></View>:
+        <View style={styles.whitePill}><Text style={styles.whitePillText}>loading...</Text></View>
+        }
+          
+          
         </LinearGradient>
       </View>
     );