+++ CHANGELOG.md
@@ -14,4 +14,5 @@
 - Add UI Catalog
 - Add Catalog List Customer w Firebase connection
 - Add 'Add New Product' Feature for Moderator
+- Add pick image function (not yet done)
 ## Fixed